storage of files was done in float32

+++ myfilemanager_sixtracklib.py
@@ -29,7 +29,7 @@ def dict_to_h5(dict_save, filename, compression_opts=4, group=None, readwrite_op
         for kk in dict_save.keys():
             if isinstance(dict_save[kk], numpy.ndarray):
                 print('Compressing '+kk)
-                dset = grp.create_dataset(kk, shape=dict_save[kk].shape, compression='gzip', compression_opts=compression_opts)
+                dset = grp.create_dataset(kk, shape=dict_save[kk].shape, dtype=dict_save[kk].dtype,compression='gzip', compression_opts=compression_opts)
                 dset[...] = dict_save[kk]
             else:
                 grp[kk] = dict_save[kk]
